1.12 Changes update.
Note these changes in this commit may vary btw updates due to using unqiue entities for generation of the nbt.

StuckToBlockPos & CollisionPos were added to all projectile entities.

Some entities in this update are only being added to the commit for improving apon the 1.11 entities nbt as shown in this 1.12 commit.

    "ExecuteOnFirstTick" & "TickDelay" tags were added to command blocks and command_block_minecarts.